# openapi.json
{
  "components": {
    "schemas": {
...
      "CalendarDate": {
        "description": "A calendar date, `YYYY-MM-DD`. Not an instant: an accounting date carries no time and no timezone, because which fiscal period an entry lands in must not depend on the reader’s.",
        "examples": [
          "2026-03-31"
        ],
        "format": "date",
        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$",
        "type": "string"
      },
      "CalendarDateInput": {
        "description": "A calendar date, `YYYY-MM-DD`. Not an instant: an accounting date carries no time and no timezone, because which fiscal period an entry lands in must not depend on the reader’s.",
        "examples": [
          "2026-03-31"
        ],
        "format": "date",
        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$",
        "type": "string"
      },
...
      "CreateFiscalPeriodRequestInput": {
        "additionalProperties": false,
        "description": "Creates one monthly period. A period is a calendar month, so it is named by year and month rather than by a date range.",
        "properties": {
          "month": {
            "description": "Calendar month, 1–12.",
            "maximum": 12,
            "minimum": 1,
            "type": "integer"
          },
          "year": {
            "description": "The calendar year the fiscal year *starts* in. A year beginning in April 2026 and ending in March 2027 is fiscal year 2026.",
            "maximum": 9998,
            "minimum": 1000,
            "type": "integer"
          }
        },
        "required": [
          "year",
          "month"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "The body of every non-2xx response.",
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "description": "Stable machine-readable failure code. Integrators branch on this, not on the HTTP status: `conflict` and `idempotency_key_conflict` are both 409 and have different recoveries. Codes are never renamed and never change status.",
                "enum": [
                  "validation_failed",
                  "unauthenticated",
                  "permission_denied",
                  "not_found",
                  "conflict",
                  "idempotency_key_conflict",
                  "precondition_failed",
                  "internal_error"
                ],
                "type": "string"
              },
              "details": {
                "additionalProperties": {},
                "propertyNames": {
                  "type": "string"
                },
                "type": "object"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "FiscalPeriod": {
        "additionalProperties": false,
        "description": "One monthly fiscal period.",
        "properties": {
          "closedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "closedByUserId": {
            "anyOf": [
              {
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "endDate": {
            "$ref": "#/components/schemas/CalendarDate"
          },
          "id": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "name": {
            "description": "Human label for the month, e.g. `2026-03`.",
            "type": "string"
          },
          "startDate": {
            "$ref": "#/components/schemas/CalendarDate"
          },
          "status": {
            "description": "A posting may only land in an `open` period (A4). Closing is the routine monthly soft close; reopening withdraws a statement that may already have been relied on, and needs a separate permission.",
            "enum": [
              "open",
              "closed"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "startDate",
          "endDate",
          "status",
          "closedAt",
          "closedByUserId"
        ],
        "type": "object"
      },
...
      "FiscalPeriodList": {
        "additionalProperties": false,
        "description": "The org’s fiscal periods in ascending date order. Unpaginated in M1.",
        "properties": {
          "periods": {
            "items": {
              "$ref": "#/components/schemas/FiscalPeriod"
            },
            "type": "array"
          }
        },
        "required": [
          "periods"
        ],
        "type": "object"
      },
...
      "GenerateFiscalYearRequestInput": {
        "additionalProperties": false,
        "description": "Generates the twelve monthly periods of one fiscal year. Generation is always explicit — no posting ever creates the period it needs (ROADMAP D-17).",
        "properties": {
          "fiscalYear": {
            "description": "The calendar year the fiscal year *starts* in. A year beginning in April 2026 and ending in March 2027 is fiscal year 2026.",
            "maximum": 9998,
            "minimum": 1000,
            "type": "integer"
          }
        },
        "required": [
          "fiscalYear"
        ],
        "type": "object"
      },
      "GeneratedFiscalYear": {
        "additionalProperties": false,
        "description": "The generated fiscal year, its span, and the twelve periods inside it.",
        "properties": {
          "endDate": {
            "$ref": "#/components/schemas/CalendarDate"
          },
          "fiscalYear": {
            "description": "The calendar year the fiscal year *starts* in. A year beginning in April 2026 and ending in March 2027 is fiscal year 2026.",
            "maximum": 9998,
            "minimum": 1000,
            "type": "integer"
          },
          "periods": {
            "description": "Twelve periods, contiguous, in ascending date order.",
            "items": {
              "$ref": "#/components/schemas/FiscalPeriod"
            },
            "type": "array"
          },
          "startDate": {
            "$ref": "#/components/schemas/CalendarDate"
          },
          "startMonth": {
            "maximum": 12,
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "fiscalYear",
          "startMonth",
          "startDate",
          "endDate",
          "periods"
        ],
        "type": "object"
      },
...
      "JournalLineRequestInput": {
        "additionalProperties": false,
        "description": "One side of one account, for a positive amount in minor units.",
        "properties": {
          "accountId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "amount": {
            "$ref": "#/components/schemas/MinorUnitsInput"
          },
          "memo": {
            "type": "string"
          },
          "side": {
            "description": "Which side of the ledger this line moves. The side carries the sign, so `amount` is always positive — a negative credit is not a debit, it is a caller that has confused the two models.",
            "enum": [
              "debit",
              "credit"
            ],
            "type": "string"
          }
        },
        "required": [
          "accountId",
          "side",
          "amount"
        ],
        "type": "object"
      },
...
      "MinorUnits": {
        "description": "An exact monetary amount as a string of minor units — an integer count of cents with no decimal point, no separators, and no leading zeros. `\"150000\"` is 1,500.00 in a two-decimal currency. Never a JSON number, and never a decimal amount: `\"1500.00\"` is refused.",
        "examples": [
          "150000",
          "-150000",
          "0"
        ],
        "pattern": "^-?(?:0|[1-9][0-9]*)$",
        "type": "string"
      },
      "MinorUnitsInput": {
        "description": "An exact monetary amount as a string of minor units — an integer count of cents with no decimal point, no separators, and no leading zeros. `\"150000\"` is 1,500.00 in a two-decimal currency. Never a JSON number, and never a decimal amount: `\"1500.00\"` is refused.",
        "examples": [
          "150000",
          "-150000",
          "0"
        ],
        "pattern": "^-?(?:0|[1-9][0-9]*)$",
        "type": "string"
      },
...
      "PostJournalRequestInput": {
        "additionalProperties": false,
        "description": "Posts one manual journal. At least two lines, and debits must equal credits exactly — there is no tolerance, because in minor units there is nothing for a tolerance to absorb.",
        "properties": {
          "date": {
            "$ref": "#/components/schemas/CalendarDateInput",
            "description": "The entry date. It must fall inside an open fiscal period — periods are never created as a side effect of posting (ROADMAP D-17), so the year has to be generated first."
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/JournalLineRequestInput"
            },
            "type": "array"
          },
          "memo": {
            "type": "string"
          }
        },
        "required": [
          "date",
          "lines"
        ],
        "type": "object"
      },
      "PostedJournal": {
        "additionalProperties": false,
        "description": "A journal as the ledger stored it, read back from the same transaction.",
        "properties": {
          "actorId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "actorType": {
            "description": "Who or what made the posting (spec §6). Taken from the session, never from the request.",
            "enum": [
              "user",
              "automation",
              "agent"
            ],
            "type": "string"
          },
          "date": {
            "$ref": "#/components/schemas/CalendarDate"
          },
          "invocationMode": {
            "anyOf": [
              {
                "enum": [
                  "interactive",
                  "scheduled"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Whether a human was present. Null means unrecorded, which is not the same as `interactive`. Set for agent callers only."
          },
          "journalId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/PostedJournalLine"
            },
            "type": "array"
          },
          "memo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "orgId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "postedAt": {
            "description": "When the journal was written. An instant, not an accounting date.",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "type": "string"
          },
          "reversesJournalId": {
            "anyOf": [
              {
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Set when this journal reverses another. The link lives on the reversing journal because the original cannot be updated (ROADMAP D-02)."
          }
        },
        "required": [
          "journalId",
          "orgId",
          "date",
          "memo",
          "postedAt",
          "actorType",
          "actorId",
          "invocationMode",
          "reversesJournalId",
          "lines"
        ],
        "type": "object"
      },
...
      "PostedJournalLine": {
        "additionalProperties": false,
        "description": "One stored line. `side` is read back from which amount column is non-zero rather than echoed from the request, so this reports what the ledger holds.",
        "properties": {
          "accountId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "amount": {
            "$ref": "#/components/schemas/MinorUnits"
          },
          "lineId": {
            "type": "string"
          },
          "memo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "side": {
            "description": "Which side of the ledger this line moves. The side carries the sign, so `amount` is always positive — a negative credit is not a debit, it is a caller that has confused the two models.",
            "enum": [
              "debit",
              "credit"
            ],
            "type": "string"
          }
        },
        "required": [
          "lineId",
          "accountId",
          "side",
          "amount",
          "memo"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Open-source, AI-native, modular double-entry accounting. Every capability in the product is reachable through this API (spec §8) — the web client is a consumer of it, not a privileged path.\n\nMonetary amounts are strings of minor units (`\"1999\"` is 19.99 in a two-decimal currency), never JSON numbers. A JSON number cannot represent a cent exactly and any client that parses one has already lost the value.\n\nEvery write endpoint requires an `Idempotency-Key` header. Reusing a key with the same request replays the original outcome; reusing it with a different request is refused with `idempotency_key_conflict`.",
    "license": {
      "identifier": "AGPL-3.0-only",
      "name": "AGPL-3.0-only"
    },
    "title": "OpenBooks API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/fiscal-periods": {
      "get": {
        "operationId": "listFiscalPeriods",
        "parameters": [
          {
            "description": "A posting may only land in an `open` period (A4). Closing is the routine monthly soft close; reopening withdraws a statement that may already have been relied on, and needs a separate permission.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "open",
                "closed"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FiscalPeriodList"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "List fiscal periods",
        "tags": [
          "fiscal-periods"
        ]
      },
      "post": {
        "description": "For the partial first year a business that started in September actually has. Gaplessness is a property of year generation, not of the table, so a year built one month at a time can leave a hole — and a date in that hole is un-postable.",
        "operationId": "createFiscalPeriod",
        "parameters": [
          {
            "description": "Required on every write (spec §12). Send one unique value per logical request and reuse it verbatim when retrying: the same key with the same request replays the original outcome, and the same key with a different request is refused with `idempotency_key_conflict`.",
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFiscalPeriodRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FiscalPeriod"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Create one monthly period",
        "tags": [
          "fiscal-periods"
        ]
      }
    },
    "/v1/fiscal-periods/{periodId}/close": {
      "post": {
        "description": "The routine monthly soft close. Records who closed it and when. A posting dated inside a closed period is refused (A4), and a posting racing this call either commits fully or not at all (A9).",
        "operationId": "closeFiscalPeriod",
        "parameters": [
          {
            "in": "path",
            "name": "periodId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "description": "Required on every write (spec §12). Send one unique value per logical request and reuse it verbatim when retrying: the same key with the same request replays the original outcome, and the same key with a different request is refused with `idempotency_key_conflict`.",
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FiscalPeriod"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Close a fiscal period",
        "tags": [
          "fiscal-periods"
        ]
      }
    },
    "/v1/fiscal-periods/{periodId}/reopen": {
      "post": {
        "description": "Withdraws a statement that may already have been relied on, so it needs the separate `periods.reopen` permission rather than the one that closes.",
        "operationId": "reopenFiscalPeriod",
        "parameters": [
          {
            "in": "path",
            "name": "periodId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "description": "Required on every write (spec §12). Send one unique value per logical request and reuse it verbatim when retrying: the same key with the same request replays the original outcome, and the same key with a different request is refused with `idempotency_key_conflict`.",
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FiscalPeriod"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Reopen a closed fiscal period",
        "tags": [
          "fiscal-periods"
        ]
      }
    },
    "/v1/fiscal-years": {
      "post": {
        "description": "The month the year begins in comes from the org’s `fiscalYearStartMonth`, not from the request — a caller that could choose it per call could generate two overlapping years for one org.",
        "operationId": "generateFiscalYear",
        "parameters": [
          {
            "description": "Required on every write (spec §12). Send one unique value per logical request and reuse it verbatim when retrying: the same key with the same request replays the original outcome, and the same key with a different request is refused with `idempotency_key_conflict`.",
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateFiscalYearRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedFiscalYear"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Generate the twelve periods of a fiscal year",
        "tags": [
          "fiscal-periods"
        ]
      }
    },
    "/v1/journals": {
      "post": {
        "description": "At least two lines, debits equal to credits exactly, every account active and in this org, and the entry date inside an open fiscal period. All of it in one transaction, so a rejected posting leaves nothing behind.",
        "operationId": "postJournal",
        "parameters": [
          {
            "description": "Required on every write (spec §12). Send one unique value per logical request and reuse it verbatim when retrying: the same key with the same request replays the original outcome, and the same key with a different request is refused with `idempotency_key_conflict`.",
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostJournalRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostedJournal"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Post a manual journal",
        "tags": [
          "journals"
        ]
      }
    },
...
  },
...
}